requirements.txt
openai>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
click>=8.1.7
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
